Tidying up existing work
Added a separate home page and login page with the home-page containing
a conditional section which changes depending on authentication status.
Added authenticated user and master template to root view class for
simple access within the template. Separated out the root factory into
collection and sample specific factories (not sure how to define these
in the route table yet - we'll get to that).