Removed cancel, finished verify complete
Removed the cancel-verification action; if it's implemented as anything
other than a no-op it opens the system to abuse (assuming the spam
recipient keeps clicking on cancel links). Better to have no method of
cancelling bogus verifications which keeps them around and slows the
creation of new ones.

Also completed implementation of the verification complete view.