Split into unit and functional tests
Split all the tests into unit-tests and functional tests (with the
awesome WebTest package). Each type has a distinct base class which
provides the fixtures (and which should enable us to select which set of
tests to run in future should one set become a bit slow at running
frequently).

Also implemented lots more tests and in the process discovered and fixed
a few bugs :)