More work on samples forms
Remembered to add the samples directory this time ... durrr.

Split validators into their own unit as it looks like we're going to
need to extend pyramid_simpleforms sufficiently that it's worth just
replicating the original code in the forms module and extending it their
(rather than sub-classing it which is just requiring more and more
complex hacks at this point).