Finished codes editing for samples
Introduced ListToDict converter to make the codes element of the form
compatible with the SA mapped collection - will probably come in handy
for the multi-role stuff in collection creation and editing too. Also
tweaked data_decoded in Form to take values from data in the case of obj
being supplied to the constructor.