Some rudimentary work on export page
Query used to get a list of the arbitrary sample codes used within the
collection, in addition to the set fields of the samples table. Still
need some JS to move fields between the lists and some additional
options (field separator, quote characters, line break characters for
CSV - not sure what for Excel - date formats for both and so on).