Tidied up helpers
Moved most of the webhelpers stuff into the view base class. Much tidier
there as some of the helpers require request-based stuff and the base
class is essentially a collection of template-rendering helpers anyway.

Also fixed up the last of the problems with anonymous viewing of open
collections (as there's no logged in user we've no idea what timezone to
convert timestamps to - use UTC until we can figure out something
better)